pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "thessla-green-modbus"
version = "2.8.0"
description = "ThesslaGreen Modbus integration for Home Assistant"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "ThesslaGreen Modbus Integration", email = "thesslagreen@users.noreply.github.com"}
]
keywords = ["home-assistant", "modbus", "hvac", "thesslagreen", "airpack"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.13",
    "Topic :: Home Automation",
]
requires-python = ">=3.13"
# Core dependencies
dependencies = ["pymodbus>=3.6.0,<4.0", "voluptuous>=0.13"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=1.2.0",
    "pytest-homeassistant-custom-component>=0.13.309,<0.13.317",
    "aiodhcpwatcher>=1.1.0",
    "aiodiscover>=2.6.1",
    "PyYAML>=6.0",
    "voluptuous>=0.13.1",
    "black>=24.4.0",
    "isort>=5.13.0",
    "ruff>=0.6.0",
    "mypy>=1.10.0",
    "pydantic>=2.0,<3",
    "pre-commit>=3.7.0",
    "types-PyYAML>=6.0.12",
]
docs = [
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=1.0.0",
]

[project.urls]
"Homepage" = "https://github.com/blakinio/thesslagreen"
"Bug Reports" = "https://github.com/blakinio/thesslagreen/issues"
"Source" = "https://github.com/blakinio/thesslagreen"
"Documentation" = "https://github.com/blakinio/thesslagreen/wiki"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["custom_components*"]

[tool.setuptools.package-data]
"custom_components.thessla_green_modbus" = [
    "options/*.json",
    "registers/*.json",
]

[tool.ruff]
line-length = 100
target-version = "py313"
src = ["custom_components", "tests", "tools"]

[tool.ruff.lint]
select = ["F", "E", "W", "B", "SIM", "RUF", "UP", "I", "PERF", "BLE"]
ignore = [
    "E501",
    "RUF002",
    "RUF003",
    "SIM102",
]
extend-safe-fixes = ["ANN"]
per-file-ignores = {"tests/*" = ["BLE001", "E402", "RUF001", "SIM117"], "tools/translate_register_descriptions.py" = ["RUF001"]}

[tool.ruff.format]
line-ending = "lf"


[tool.black]
target-version = ["py313"]
line-length = 100
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.13"
files = ["custom_components/thessla_green_modbus"]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
namespace_packages = true
explicit_package_bases = true
mypy_path = ["."]
follow_imports = "skip"

[[tool.mypy.overrides]]
module = [
    "homeassistant.*",
    "pymodbus.*",
    "pydantic.*",
    "voluptuous.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tests.*"]
ignore_errors = true

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = [".git", "testing_config"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--disable-warnings",
    "--tb=short",
    "-ra",
    "-q",
]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "asyncio: marks asynchronous tests",
]


[tool.coverage.run]
source = ["custom_components.thessla_green_modbus"]
branch = true
omit = [
    "custom_components/thessla_green_modbus/_compat.py",
    "custom_components/thessla_green_modbus/translations/*",
]

[tool.coverage.report]
fail_under = 80
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
